Refactor mock hierarchy
