+++ lib/src/Mock/Components/ComponentScreen.tsx
@@ -3,7 +3,7 @@ import { Button, View, Text } from 'react-native';
 import { Navigation } from '../..';
 import { ComponentProps } from '../ComponentProps';
 import { VISIBLE_SCREEN_TEST_ID } from '../constants';
-import { LayoutStore } from '../LayoutStore';
+import { LayoutStore } from '../Stores/LayoutStore';
 import { connect } from '../connect';
 import { TopBar } from './TopBar';
