+++ lib/src/Mock/Components/Overlays.tsx
@@ -1,17 +1,18 @@
 import React, { Component } from 'react';
 import { View } from 'react-native';
-import { ComponentProps } from './ComponentProps';
+import { ComponentProps } from '../ComponentProps';
+import { VISIBLE_OVERLAY_TEST_ID } from '../constants';
 import { LayoutComponent } from './LayoutComponent';
-import ParentNode from './Layouts/ParentNode';
-import { LayoutStore } from './LayoutStore';
-import { connect } from './connect';
+import ParentNode from '../Layouts/ParentNode';
+import { LayoutStore } from '../Stores/LayoutStore';
+import { connect } from '../connect';
 
-export const Modals = connect(
+export const Overlays = connect(
   class extends Component<ComponentProps> {
     render() {
-      const children = LayoutStore.getModals();
+      const children = LayoutStore.getOverlays();
       return (
-        <View testID={'MODALS'}>
+        <View testID={VISIBLE_OVERLAY_TEST_ID}>
           {children.map((child: ParentNode) => {
             return <LayoutComponent key={child.nodeId} layoutNode={child} />;
           })}
