+++ lib/src/Mock/Components/TopBar.tsx
@@ -4,8 +4,8 @@ import { Navigation, OptionsTopBarButton } from '../..';
 import { OptionsTopBar, OptionsTopBarBackButton } from '../../interfaces/Options';
 import { DEFAULT_BACK_BUTTON_TEST_ID } from '../constants';
 import ParentNode from '../Layouts/ParentNode';
-import { LayoutStore } from '../LayoutStore';
-import { NavigationButton } from '../NavigationButton';
+import { LayoutStore } from '../Stores/LayoutStore';
+import { NavigationButton } from './NavigationButton';
 
 export interface TopBarProps {
   layoutNode: ParentNode;
