+++ lib/src/Mock/Layouts/ParentNode.ts
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import { OptionsTopBarButton } from '../../interfaces/Options';
 import { Options } from '../../index';
-import { LayoutStore } from '../LayoutStore';
+import { LayoutStore } from '../Stores/LayoutStore';
 import ComponentNode from './ComponentNode';
 import LayoutNodeFactory from './LayoutNodeFactory';
 import Node, { NodeType } from './Node';
