+++ lib/src/Mock/index.ts
@@ -1,4 +1,4 @@
-jest.mock('./../adapters/NativeCommandsSender', () => require('./NativeCommandsSender.tsx'));
-jest.mock('./../adapters/NativeEventsReceiver', () => require('./NativeEventsReceiver.ts'));
+jest.mock('./../adapters/NativeCommandsSender', () => require('./mocks/NativeCommandsSender.tsx'));
+jest.mock('./../adapters/NativeEventsReceiver', () => require('./mocks/NativeEventsReceiver.ts'));
 
 export const Application = require('./Application').Application;
